refactor: API 로직 리팩토링 작업(User-try.catch)
