+++ src/api/Lecture.js
@@ -3,32 +3,17 @@ import JwtInterceptors from './ApiController';
 const Lecture = () => {
   const instance = JwtInterceptors().instance;
 
+  // 메인페이지
   const main = async (lecture, page, majorType) => {
-    return instance({
-      url: `/lecture/all/?option=${lecture}&page=${page}&majorType=${majorType}`,
-      method: 'GET',
-    });
+    return instance.get(`/lecture/all/?option=${lecture}&page=${page}&majorType=${majorType}`);
   };
 
-  const mainInfinite = async (lecture, pageParam, majorType) => {
-    const result = await instance({
-      url: `/lecture/all/?option=${lecture}&page=${pageParam}&majorType=${majorType}`,
-      method: 'GET',
-    });
-    return {
-      data: result,
-      isLast: result.data.length < 10,
-      nextPage: pageParam + 1,
-    };
-  };
-
-  // 통합검색결과Api
+  // 통합검색결과
   //꿀강순[modifiedDate, lectureSatisfactionAvg, lectureHoneyAvg, lectureLearningAvg]
   const search = async (searchValue, pageParam, option, major) => {
-    const result = await instance({
-      url: `/lecture/search/?searchValue=${searchValue}&option=${option}&page=${pageParam}&majorType=${major}`,
-      method: 'GET',
-    });
+    const result = await instance.get(
+      `/lecture/search/?searchValue=${searchValue}&option=${option}&page=${pageParam}&majorType=${major}`
+    );
     return {
       data: result,
       isLast: result.data.length < 10,
@@ -38,18 +23,12 @@ const Lecture = () => {
 
   // 검색 결과 자세히보기 (Lecture)
   const detail = (selectId) => {
-    return instance({
-      url: `/lecture/?lectureId=${selectId}`,
-      method: 'GET',
-    });
+    return instance.get(`/lecture/?lectureId=${selectId}`);
   };
 
   // 검색 결과 자세히보기 (Evaluation)
   const evaluation = async (selectId, pageParam) => {
-    const result = await instance({
-      url: `/evaluate-posts/?lectureId=${selectId}&page=${pageParam}`,
-      method: 'GET',
-    });
+    const result = await instance.get(`/evaluate-posts/?lectureId=${selectId}&page=${pageParam}`);
     return {
       data: result,
       isLast: result.data.length < 10,
@@ -59,10 +38,7 @@ const Lecture = () => {
 
   // 검색 결과 자세히보기 (Exam)
   const examInfo = async (selectId, pageParam) => {
-    const result = await instance({
-      url: `/exam-posts/?lectureId=${selectId}&page=${pageParam}`,
-      method: 'GET',
-    });
+    const result = await instance.get(`/exam-posts/?lectureId=${selectId}&page=${pageParam}`);
     return {
       data: result,
       isLast: result.data.length < 10,
@@ -70,7 +46,7 @@ const Lecture = () => {
     };
   };
 
-  return { main, mainInfinite, search, evaluation, examInfo, detail };
+  return { main, search, evaluation, examInfo, detail };
 };
 
 export default Lecture;
