+++ src/constants/cacheTime.js
@@ -1,0 +1,3 @@
+export const CACHE_TIME = {
+  MINUTE_30: 1000 * 60 * 30,
+};
